feat(pagination): support customizable-per-type max + default page sizes
Max page size is the largest value that a client is allowed to specify in page[limit] before getting an error. Default page size is used when the client doesn't specify a page[limit] at all. For backwards compatibility, both effectively default to infinity.

One open question with this PR was where to validate the end-user-provided page size against the limit set by the API author:

- In the adapter? That would best guarantee that the limit is respected, but it would: a) require passing the per-type pagination limits to the adapter on construction and b) might inadvertently prevent custom-constructed queries (the results of which may never even be shown to the user) from using a higher limit.

- When constructing the default query? This would avoid downside a) and b) above, but might lead to users forgetting to apply the pagination limits when constructing custom queries [the flip-side of b) above].

For safety, the approach I settled on was to validate the limit in the adapter (eventually, the library can eventually expose a function for this for all adapters to share), so that custom queries can't inadvertently forget to set it. Then, I created a field on `Query` (`ignoreLimitMax`) that allows custom-constructed queries to explicitly indicate they want to use/allow a bigger `limit`.

The one tricky bit with this approach is that it still required getting the per-type pagination settings into the adapter somehow, and there's a bit of a cyclic dependency issue with that (resource descriptions are typically created before the registry instance and point to the adapter instance, but that adapter instance would need to have access to the pagination settings from the type descriptions). So, I worked around this by adding a required `setRegistryDerivedOptions` method to the adapter interface that the registry automatically calls when it's given the resource type descriptions.